refactor home view code and imports
