+++ lib/config/index.dart
@@ -1,3 +1,4 @@
 export './router/index.dart';
 export './themes/index.dart';
 export './constants/enviroment.dart';
+export './plugins/index.dart';
