+++ lib/config/plugins/index.dart
@@ -1,0 +1,2 @@
+export './auto_size_text.adapter.dart';
+export './carousel_slider.adapter.dart';
