+++ lib/presentation/widgets/view/home_view/index.dart
@@ -1,0 +1,4 @@
+export './carousel_welcom_box.dart';
+export './welcome_box.dart';
+export './custom_box_date.dart';
+export './available_dishes_carousel.dart';
