+++ lib/presentation/widgets/view/index.dart
@@ -1,1 +1,2 @@
 export './custom_container_food.dart';
+export './home_view/index.dart';
